requirements.txt
Django
djangorestframework
python-dotenv
PyPDF2
requests
elevenlabs
psycopg